moves pipeline tests to pipelines
